app/requirements.txt
boto3==1.12.44
botocore==1.15.44
click==7.1.1
docutils==0.15.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
MarkupSafe==1.1.1
mysql-connector==2.2.9
numpy==1.18.3
opencv-python==4.2.0.34
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
